Rename async query package
